requirements_O.txt
numpy
scipy
pandas
scikit-learn
matplotlib
tqdm
rich
torchtext==0.13.1
lit # Getting installation errors with torch 2.0 if this isn't installed
pytorch-lightning==1.9.3
hydra-core
omegaconf
wandb
einops
cmake # For pykeops support
# pykeops # Seems to cause various issues; leaving uninstalled by default
transformers # For some schedulers

# Model specific packges
# pytorch-fast-transformers # for Performer

# Dataset specific packages
datasets # LRA
sktime # BIDMC
scikit-learn # Impedance
numba # Impedance
gluonts # Monash
timm==0.5.4 # ImageNet